+++ packages/discovery-ui/types/ShaclRenderer.d.ts
@@ -1,0 +1,6 @@
+type ShaclRendererProps = {
+    readonly value: string;
+};
+export declare function ShaclRenderer({ value }: ShaclRendererProps): JSX.Element;
+export {};
+//# sourceMappingURL=ShaclRenderer.d.ts.map