+++ packages/icons-ui/types/ScreenLayout.d.ts
@@ -1,0 +1,6 @@
+export type ScreenLayoutProps = {
+    readonly children: React.ReactNode;
+    readonly header?: React.ReactNode;
+};
+export declare function ScreenLayout({ header, children, }: ScreenLayoutProps): JSX.Element;
+//# sourceMappingURL=ScreenLayout.d.ts.map