+++ packages/icons-ui/types/stories/components.stories.d.ts
@@ -1,0 +1,5 @@
+import { Meta } from "@storybook/react";
+declare const _default: Meta;
+export default _default;
+export declare const Overview: () => import("react/jsx-runtime").JSX.Element;
+//# sourceMappingURL=components.stories.d.ts.map