+++ packages/nextjs-ui/types/PagerForm.d.ts
@@ -1,0 +1,4 @@
+import { FieldValues } from "react-hook-form";
+import { PagerFormProps } from "./PagerForm.types";
+export declare function PagerForm<TFieldValues extends FieldValues>({ ...props }: PagerFormProps<TFieldValues>): JSX.Element;
+//# sourceMappingURL=PagerForm.d.ts.map