+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @truststack/ui
 
+## 0.5.0
+
+### Minor Changes
+
+- e834c29: Bumping tamagui version.
+
+### Patch Changes
+
+- Updated dependencies [e834c29]
+  - @truststack/icons-ui@0.5.0
+
 ## 0.4.1
 
 ### Patch Changes